+++ src/nuself/trace/__init__.py
@@ -1,0 +1,23 @@
+"""Trace subsystem public interfaces."""
+
+from nuself.trace.domain import (
+    ThoughtTrace,
+    TraceKind,
+    TraceLink,
+    TraceRelation,
+    TraceVisibility,
+)
+from nuself.trace.repository import TraceNotFound, TraceRepository
+from nuself.trace.service import TraceQueryService, TraceRecorder
+
+__all__ = [
+    "ThoughtTrace",
+    "TraceKind",
+    "TraceLink",
+    "TraceNotFound",
+    "TraceQueryService",
+    "TraceRecorder",
+    "TraceRelation",
+    "TraceRepository",
+    "TraceVisibility",
+]